requirements.txt
python-dotenv
pandas
sqlalchemy
numpy
